Extract TestingValuesWriter from TestDefinitionLevels
